fix: add types location for cjs/src/
Now we're publishing a `"main"`, TypeScript is using that rather than the
`"exports"` map. It points to the CJS published version, located in cjs/src
in the bundle. This adds a descriptor that tells TypeScript to look for types
for that location in the types/, which should be the same as to what it finds
in the export map.

Ref: https://github.com/multiformats/js-multiformats/pull/64#issuecomment-792393859