feat!: remove legacy function
The legacy function in this module does not return objects that are
compatible with the types from `interface-ipld-format`, worse, the
way to make them compatible is to include references to the types
from the codec name/number tables etc from the modules that this
module is supposed to replace.

There's now the `blockcodec-to-ipld-format` module that does the
icky job of converting from block codecs to ipld formats so can
depend on bits of the legacy stack so this module doesn't have to.

This PR removes the legacy function and the deps it uses, people
should use `blockcodec-to-ipld-format` instead to get ts compatible
ipld formats and this module can be legacy (and dep) free for the
future.