feat: support v0 identity cids
When dealing with PeerIDs we commonly encode them as v0 CIDs using
base58btc encoded multihashes with the identtiy hash.

The change here is to support `1` prefixed CIDv0 as a special case
(e.g. the identity hash) similar to how we support `Q` prefixed
CIDv0 as a special case (e.g. sha2-256).

Fixes #90