fix: add .or method to types
The `Decoder` class implements a `.or` method but it's not included
in any of the interfaces it implements (MultibaseDecoder, UnibaseDecoder
or BaseDecoder).

I'm not sure if it belongs on the MultibaseDecoder interface, but
the things that are exported have `prefix`, `name`, `encode` and `decode`
props which make them seem like `MultibaseCodec`s and the `decode`
prop of a `MultibaseCodec` is a `MultibaseDecoder` so that's why I
put the `.or` defintion there.