feat: formally expose MultihashHasher#encode
It's always been there, and gets defined by the class, just not in the
interface. SyncMultihashHasher hasn't had it, but in practice they all
do have it (unless someone else has authored one since it was introduced
that we don't know about).

The addition to SyncMultihashHasher is technically breaking since it adds an
additional burden to the API for implementers. But it's both very new, and the
implementations (we're aware of) already implement with `encode()`.