fix: remove use of Object.defineProperties in CID class
`Object.defineProperties` is a performance bottleneck in applications that
create lots and lots of CIDs (e.g. IPFS) so this PR removes it.

The `asCID` property is changed to be a [private class field](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Classes/Private_class_fields)
which requires increasing the minimum supported EcmaScript version but I don't
know if that's a big deal or not.  It does seem to make the property non-enumerable
though.

The CID class now implements a `Link` interface that has public `byteOffset`
and `byteLength` properties so these become regular properties

`code`, `version`, `multihash` and `bytes` become writable/configurable
but they are marked with `@readonly` so maybe that's enough?

Fixes #200