fix: publish typedocs in gh-pages branch
Updates aegir to the latest version which has reinstated the `docs`
command to publish typedocs in the `gh-pages` branch of repos.

A `typedoc-urls.json` file is now generated as part of the docs run
which allows cross-linking between module docs.

This file lives at the "well known" location of `dist/typedoc-urls.json`
in the published tarball, so the `files` list in `package.json` needed
changing as it's currently only publishing `dist/types`.